Modfied p type for Zadacha 4

+++ Zadacha4/Program.cs
@@ -1,7 +1,7 @@
 ﻿// Написать программу вычисления произведения чисел от 1 до N
 Console.Write("Введите число n: ");
 int n = int.Parse(Console.ReadLine() ?? "0");
-int p = 1;
+double p = 1;
 for (int i = 1; i <= n; i++)
 {
     p = p * i;
